feat(formula): get dirty data
